FIXED ENGLISH. See full commit comment below for details.
We fix an error in the original English, identified by our translators.
In the MC160.dev.17 story, in the 3rd question, the correct answer is B, but the answer text incorrectly says "pink" flowers, instead of "yellow" flowers.
Our translations for Mauritian and Haitian-1 and Haitian-2 have already been corrected by the translators, to say "yellow" flowers.
Here, we fix the English.